server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
numpy>=2.0.0
pandas>=2.2.0
opencv-python>=4.10.0
scikit-learn>=1.5.0
Pillow>=11.0.0
requests>=2.31.0
